Add and update pooling tests
