+++ tests/Makefile
@@ -2,7 +2,7 @@ TESTS=add add_2 add_3 \
 	  argmax const matmul linreg placeholder relu \
 	  min_1 min_1_1 min_2 min_3 min_3_1 \
 	  max_1_1 max_1 max_2 max_3 max_3_1 \
-		max_pool_1 \
+	  max_pool_1 max_pool_2 \
 	  reshape_2 reshape_3 reshape_4 \
 	  reshape_5 reshape_6 reshape_7 \
 	  conv
