+++ tests/max_pool_2/main.cpp
@@ -1,4 +1,4 @@
-#include "max_pool_1_ctx.hpp"
+#include "max_pool_2_ctx.hpp"
 #include "uTensor/loaders/tensorIdxImporter.hpp"
 #include "uTensor/util/uTensor_util.hpp"
 #include "TESTS/test.hpp"
@@ -7,7 +7,7 @@
 #include <SDBlockDevice.h>
 
 
-class MaxPoolTest1 : public Test {
+class MaxPoolTest2 : public Test {
     Context ctx;
     TensorIdxImporter t_import;
 public:
@@ -24,7 +24,7 @@ int main(int argc, char* argv[]) {
     ON_ERR(bd.init(), "SDBlockDevice init ");
     ON_ERR(fs.mount(&bd), "Mounting the filesystem on \"/fs\". ");
 
-    MaxPoolTest1 test;
+    MaxPoolTest2 test;
     test.runAll();
     test.printSummary();
 
@@ -34,18 +34,18 @@ int main(int argc, char* argv[]) {
     return 0;
 }
 
-void MaxPoolTest1::runAll(void) {
-    testStart("simple max_pool_1 test");
+void MaxPoolTest2::runAll(void) {
+    testStart("simple max_pool_2 test");
     timer_start();
-    get_test_quant_max_pool_1_ctx(ctx);
-    S_TENSOR pool1 = ctx.get("pool1:0");
+    get_test_quant_max_pool_2_ctx(ctx);
+    S_TENSOR pool2 = ctx.get("pool2:0");
     ctx.eval();
     timer_stop();
 
-    Tensor* ref_pool = t_import.float_import("/fs/idx_data/output_pool.idx");
+    Tensor* ref_pool2 = t_import.float_import("/fs/idx_data/output_max_pool_2.idx");
 
     // compare the results
-    double err = meanPercentErr<float>(ref_pool, pool1.get());
+    double err = meanPercentErr<float>(ref_pool2, pool2.get());
     printf("meanPercentErr: %f (<0.03?)\n", err);
     passed(err < 0.03);
 }
