+++ tests/max_pool_2/output_max_pool_2.idx
